requirements.txt
streamlit
tensorflow
Pillow
numpy
matplotlib
requests
google-generativeai
